http: Expose connection_factory implementations
Websocket protocol handshake uses HTTP, thus it makes sense to reuse at
least some of the abstractions and API style.
