+++ src/http/client.cc
@@ -35,7 +35,6 @@ module seastar;
 #include <seastar/core/loop.hh>
 #include <seastar/core/when_all.hh>
 #include <seastar/core/reactor.hh>
-#include <seastar/net/tls.hh>
 #include <seastar/http/client.hh>
 #include <seastar/http/request.hh>
 #include <seastar/http/reply.hh>
@@ -203,38 +202,10 @@ future<> connection::close() {
     });
 }
 
-class basic_connection_factory : public connection_factory {
-    socket_address _addr;
-public:
-    explicit basic_connection_factory(socket_address addr)
-            : _addr(std::move(addr))
-    {
-    }
-    virtual future<connected_socket> make(abort_source* as) override {
-        return seastar::connect(_addr, {}, transport::TCP);
-    }
-};
-
 client::client(socket_address addr)
         : client(std::make_unique<basic_connection_factory>(std::move(addr)))
 {
 }
-
-class tls_connection_factory : public connection_factory {
-    socket_address _addr;
-    shared_ptr<tls::certificate_credentials> _creds;
-    sstring _host;
-public:
-    tls_connection_factory(socket_address addr, shared_ptr<tls::certificate_credentials> creds, sstring host)
-            : _addr(std::move(addr))
-            , _creds(std::move(creds))
-            , _host(std::move(host))
-    {
-    }
-    virtual future<connected_socket> make(abort_source* as) override {
-        return tls::connect(_creds, _addr, tls::tls_options{.server_name = _host});
-    }
-};
 
 client::client(socket_address addr, shared_ptr<tls::certificate_credentials> creds, sstring host)
         : client(std::make_unique<tls_connection_factory>(std::move(addr), std::move(creds), std::move(host)))
